refactor(aegis): slim rpc/hermes to only expose upstream-dependent methods
- Rename hermesclient/ -> rpc/hermes/, package hermes
- Reorganize by gRPC service: provision.go, resource.go, key.go, user.go
- Remove all management CRUD not needed by aegis runtime:
  - provision: remove Create/List/Update Application, Service, Domain CRUD
  - provision: remove DomainIDPConfig CRUD, ApplicationIDPConfig CUD
  - resource: remove all Relationship write ops, AppService write ops
  - key: remove RotateKey, IDP Key CRUD (kept ResolveIDPKey)
  - user: remove GetByOpenID, GetByIdentity, Group CRUD
- Keep only methods required by contract.HermesProvider, UserProvider,
  iris.CredentialStore, and idp.KeyResolver interfaces
- Remove unused convert helpers (safeInt8, safeInt32, safeUint32, etc)
- Net reduction: ~795 lines
